feat: add toJson and toYaml
