+++ packages/openapi-parser/src/utils/openapi/dereferenceCommand.ts
@@ -2,6 +2,8 @@ import type { DereferenceResult, Queue, Task } from '../../types'
 import type { DereferenceOptions } from '../dereference'
 import type { ValidateOptions } from '../validate'
 import { get } from './get'
+import { toJson } from './toJson'
+import { toYaml } from './toYaml'
 import { upgradeCommand } from './upgradeCommand'
 import { queueTask } from './utils/queueTask'
 import { validateCommand } from './validateCommand'
@@ -39,6 +41,8 @@ export function dereferenceCommand<T extends Task[]>(
   return {
     get: () => get(queue),
     upgrade: () => upgradeCommand(queue),
+    toJson: () => toJson(queue),
+    toYaml: () => toYaml(queue),
     validate: (validateOptions?: ValidateOptions) =>
       validateCommand(queue, validateOptions),
   }
