+++ packages/openapi-parser/src/utils/openapi/get.ts
@@ -1,14 +1,5 @@
-import type {
-  CommandChain,
-  Merge,
-  PromiseReturnType,
-  Queue,
-  Task,
-} from '../../types'
-import { dereference } from '../dereference'
-import { load } from '../load'
-import { upgrade } from '../upgrade'
-import { validate } from '../validate'
+import type { CommandChain, Queue, Task } from '../../types'
+import { workThroughQueue } from './utils/workThroughQueue'
 
 /**
  * Run the chained tasks and return the results
@@ -16,50 +7,5 @@ import { validate } from '../validate'
 export async function get<T extends Task[]>(
   queue: Queue<T>,
 ): Promise<CommandChain<T>> {
-  let result = {} as CommandChain<T>
-
-  // Work through the whole queue
-  queue.tasks.forEach(async ({ name }: { name: keyof Commands }) => {
-    const { input } = queue
-
-    // load
-    if (name === 'load') {
-      result = {
-        ...result,
-        ...(await load(input)),
-      } as Merge<typeof result, PromiseReturnType<typeof load>>
-    }
-
-    // upgrade
-    else if (name === 'dereference') {
-      result = {
-        ...result,
-        ...(await dereference(input)),
-      } as Merge<typeof result, PromiseReturnType<typeof dereference>>
-    }
-
-    // upgrade
-    else if (name === 'upgrade') {
-      result = {
-        ...result,
-        ...upgrade(input),
-      } as Merge<typeof result, ReturnType<typeof upgrade>>
-    }
-
-    // validate
-    else if (name === 'validate') {
-      result = {
-        ...result,
-        ...validate(input),
-      } as Merge<typeof result, ReturnType<typeof validate>>
-    }
-
-    // Make TS complain when we forgot to handle a command.
-    else {
-      // eslint-disable-next-line @typescript-eslint/no-unused-vars
-      const _: never = name
-    }
-  })
-
-  return result
+  return await workThroughQueue(queue)
 }
