+++ packages/openapi-parser/src/utils/openapi/validateCommand.ts
@@ -3,6 +3,8 @@ import type { DereferenceOptions } from '../dereference'
 import type { ValidateOptions } from '../validate'
 import { dereferenceCommand } from './dereferenceCommand'
 import { get } from './get'
+import { toJson } from './toJson'
+import { toYaml } from './toYaml'
 import { upgradeCommand } from './upgradeCommand'
 import { queueTask } from './utils/queueTask'
 
@@ -40,6 +42,8 @@ export function validateCommand<T extends Task[]>(
     dereference: (dereferenceOptions?: DereferenceOptions) =>
       dereferenceCommand(queue, dereferenceOptions),
     get: () => get(queue),
+    toJson: () => toJson(queue),
+    toYaml: () => toYaml(queue),
     upgrade: () => upgradeCommand(queue),
   }
 }
